requirements.txt
Jinja2==3.0.3
pika==1.2.0
pydantic==1.9.0
psycopg2-binary==2.9.2
requests==2.27.1
fastapi==0.61.1
uvicorn==0.12.2
orjson==3.4.1
websockets==10.2.0
